+++ settings.php
@@ -38,6 +38,14 @@ if ($hassiteconfig) {
     $setting = new admin_setting_configtextarea($name, $title, $description, $default);
     $settings->add($setting);
 
+    // Require the user to be logged-in in order to send the form.
+    $default = 0;
+    $name = 'local_contact/loginrequired';
+    $title = get_string('loginrequired', 'local_contact');
+    $description = get_string('loginrequired_description', 'local_contact');
+    $setting = new admin_setting_configcheckbox($name, $title, $description, $default);
+    $settings->add($setting);
+
     // Override and disable ReCAPTCHA, if the private and public keys are setup in Moodle.
     if (!empty($CFG->recaptchaprivatekey) AND !empty($CFG->recaptchapublickey)) {
         // Information on using recaptcha with Contact Form.